requirements.txt
streamlit
langdetect
googletrans==4.0.0-rc1
transformers
torch
duckduckgo-search
Pillow
